Extract component CookieContent from CookieBanner.tsx
